+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/inspections/cleanup/RobotUnusedKeywordOnlyMarkerInspection.java
@@ -15,6 +15,7 @@ import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotLocalArgumentsS
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotLocalArgumentsSettingParameterOptional;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotVariableDefinition;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotVisitor;
+import dev.xeonkryptos.xeonrobotframeworkplugin.quickfix.RemoveUnusedKeywordOnlyMarkerQuickFix;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Set;