+++ src/main/kotlin/dev/xeonkryptos/xeonrobotframeworkplugin/quickfix/ReplaceRunKeywordAndReturnIfQuickFix.kt
@@ -1,4 +1,4 @@
-package dev.xeonkryptos.xeonrobotframeworkplugin.inspections.maintainability
+package dev.xeonkryptos.xeonrobotframeworkplugin.quickfix
 
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiFile
@@ -9,28 +9,30 @@ import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotVisitor
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.util.RobotElementGenerator
 import dev.xeonkryptos.xeonrobotframeworkplugin.util.RobotNames
 
-class ReplaceContinueForLoopIfQuickFix(keywordCall: RobotKeywordCall) : AbstractReplaceKeywordQuickFix(keywordCall, RobotNames.CONTINUE_FOR_LOOP_IF_NORMALIZED_KEYWORD_NAME) {
+class ReplaceRunKeywordAndReturnIfQuickFix(keywordCall: RobotKeywordCall) : AbstractReplaceKeywordQuickFix(keywordCall, RobotNames.RUN_KEYWORD_AND_RETURN_IF_NORMALIZED_KEYWORD_NAME) {
 
-    override fun isAvailable(keywordCall: RobotKeywordCall): Boolean {
-        val argumentsCollector = QuickFixArgumentsCollector()
-        keywordCall.positionalArgumentList.forEach { it.accept(argumentsCollector) }
-        return argumentsCollector.conditionalContent != null
-    }
+    override fun isAvailable(keywordCall: RobotKeywordCall): Boolean = keywordCall.positionalArgumentList.isNotEmpty()
 
     override fun invoke(project: Project, file: PsiFile, keywordCall: RobotKeywordCall) {
         val argumentsCollector = QuickFixArgumentsCollector()
         keywordCall.positionalArgumentList.forEach { it.accept(argumentsCollector) }
 
         val conditionalContent = argumentsCollector.conditionalContent?.text ?: return
+        val subKeywordCall = argumentsCollector.subKeywordCall?.text ?: return
 
-        val loopControlStructure =
-            RobotElementGenerator.getInstance(project).createNewConditionalStructure(conditionalContent, RobotElementGenerator.LoopControlStructureType.CONTINUE.name, emptyArray(), null)
-        keywordCall.replace(loopControlStructure)
+        val conditionalBodyWithReturn = """
+            ${'$'}{Result}=  $subKeywordCall
+            ${RobotNames.RETURN_RESERVED_NAME}  ${'$'}{Result}
+        """.trimIndent()
+        val conditionalStructure = RobotElementGenerator.getInstance(project).createNewConditionalStructure(conditionalContent, conditionalBodyWithReturn, emptyArray(), null)
+
+        keywordCall.replace(conditionalStructure)
     }
 
     private class QuickFixArgumentsCollector : RobotVisitor() {
 
         var conditionalContent: RobotConditionalContent? = null
+        var subKeywordCall: RobotKeywordCall? = null
 
         override fun visitPositionalArgument(o: RobotPositionalArgument) {
             o.acceptChildren(this)
@@ -38,6 +40,10 @@ class ReplaceContinueForLoopIfQuickFix(keywordCall: RobotKeywordCall) : Abstract
 
         override fun visitConditionalContent(o: RobotConditionalContent) {
             conditionalContent = o
+        }
+
+        override fun visitKeywordCall(o: RobotKeywordCall) {
+            subKeywordCall = o
         }
     }
 }